Banking service updated with account

+++ banking-web-services/src/main/java/com/yash/banking/webervices/banking_web_services/repository/AccountTypeRepository.java
@@ -2,10 +2,11 @@ package com.yash.banking.webervices.banking_web_services.repository;
 
 import com.yash.banking.webervices.banking_web_services.dto.AccountTypes;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 @Repository
-public interface AccountTypeRepository extends JpaRepository<AccountTypes, Integer> {
+public interface AccountTypeRepository extends JpaRepository<AccountTypes, Long> {
 
 }
